Tell the agent to kill a job if it didn't start properly

+++ agent/internal/hashcat/hashcat.go
@@ -307,6 +307,9 @@ func (sess *HashcatSession) Start() error {
 }
 
 func (sess *HashcatSession) Kill() error {
+	if sess.proc == nil || sess.proc.Process == nil {
+		return nil
+	}
 	return sess.proc.Process.Kill()
 }
